sentence_explainer: prefer larger phrases for brevity
This seems to help a little bit. Would be nice to have some better evals to see the effects.

I noticed the gemini-2.0-flash model slipped in some Russian in its explanations a couple times.

And it also added some chit-chat sentences before and after the sentence breakdown. Trying to refine the prompt to discourage that.

Eventually, the final solution will probably just be to generate a schema-enforced response.
And can make a nice UI to display that, like with clickable phrases in the sentence to see the details.